all new league_db generator
this one creates parquet files from data pulled from the data API.  much of the data is nested which will require dbs like duckdb or postgres to be able to query efficiently.

however it is much better for keeping a 1:1 relationship with how the data is presented by the raw API so more resilient to changes and more consistent.